refactor: move rsync server test util
