+++ cmd/dump/rsync_delta_integration_test.go
@@ -16,8 +16,8 @@ import (
 	"lvmsync_go/device"
 	"lvmsync_go/internal/config"
 	digestpkg "lvmsync_go/internal/digest"
-	"lvmsync_go/internal/rsyncserver"
 	"lvmsync_go/internal/rsyncwire"
+	"lvmsync_go/testutil/rsyncserver"
 )
 
 type countingConn struct {
